Moved compressionhelper to /Encoding folder and changed gzip -> LZ4
